requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-search-documents==11.5.2
backoff==2.2.1
banks==2.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.8.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.1.1
griffe==1.7.1
grpcio==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==8.34.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
llama-cloud==0.1.17
llama-cloud-services==0.6.9
llama-index==0.12.26
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.27
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-llms-gemini==0.4.14
llama-index-llms-google-genai==0.1.7
llama-index-llms-openai==0.3.29
llama-index-memory-mem0==0.3.0
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.4.post1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mem0ai==0.1.79
monotonic==1.6
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==1.32.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.69.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.7
portalocker==2.10.1
posthog==3.23.0
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.3.0
qdrant-client==1.13.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
rsa==4.9
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
streamlit==1.44.0
striprtf==0.0.26
tenacity==9.0.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarl==1.18.3
